requirements.txt
pymysql
flask_sqlalchemy
python-dotenv
